refactor(core): split config.rs into settings and device submodules
Pure code motion: the settings/value types (AppSettings, Appearance,
Lighting, WheelMode, SmartShift, GestureOwner) and their serde helpers
move to config/settings.rs; DeviceIdentity, DeviceConfig, and the
RawDeviceConfig migration shim move to config/device.rs. The root
config.rs keeps Config, ConfigError, the runtime operations, and
write_atomic, and re-exports every previously public item so all
openlogi_core::config:: paths are unchanged.
