+++ crates/openlogi-core/src/config.rs
@@ -15,8 +15,17 @@ use std::{
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
+mod device;
+mod settings;
+
+pub use device::{DeviceConfig, DeviceIdentity};
+pub use settings::{
+    AppSettings, Appearance, DEFAULT_THUMBWHEEL_SENSITIVITY, GestureOwner, Lighting,
+    MAX_THUMBWHEEL_SENSITIVITY, MIN_THUMBWHEEL_SENSITIVITY, SMARTSHIFT_AUTO_DISENGAGE_DEFAULT,
+    SMARTSHIFT_MIN_AUTO_DISENGAGE, SmartShift, WheelMode,
+};
+
 use crate::binding::{Action, Binding, ButtonId, GestureDirection, default_binding_for};
-use crate::device::{Capabilities, DeviceKind, DeviceModelInfo};
 use crate::paths::{self, PathsError};
 
 /// The schema version the current build produces. Bumped on breaking layout
@@ -57,512 +66,6 @@ impl Default for Config {
             app_settings: AppSettings::default(),
             selected_device: None,
             devices: BTreeMap::new(),
-        }
-    }
-}
-
-/// Light/dark appearance preference. `System` follows the OS appearance (the
-/// historical behaviour); `Light` / `Dark` force a mode regardless of the OS.
-/// Platform-free so the core crate stays GUI-agnostic — the GUI maps this onto
-/// gpui-component's `ThemeMode`.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum Appearance {
-    /// Follow the operating system's light/dark setting.
-    #[default]
-    System,
-    /// Always use the light variant of the selected theme.
-    Light,
-    /// Always use the dark variant of the selected theme.
-    Dark,
-}
-
-/// App-wide preferences not tied to any particular device.
-///
-/// All fields are `#[serde(default)]` so adding a new one is backward
-/// compatible — old config files just keep the default for the new field.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[allow(
-    clippy::struct_excessive_bools,
-    reason = "independent on/off user preferences, not a state machine"
-)]
-pub struct AppSettings {
-    /// When true, a macOS `LaunchAgent` plist at
-    /// `~/Library/LaunchAgents/org.openlogi.openlogi.plist` is installed
-    /// so the app starts on login (P2.2). The plist is reconciled with
-    /// this field on every startup; flipping the flag and relaunching is
-    /// enough to install / remove it.
-    #[serde(default)]
-    pub launch_at_login: bool,
-    /// Opt-in update check (P2.8). **Off by default** to honour the
-    /// README's "no telemetry, no auto-update poller" promise. When true,
-    /// the app makes exactly one `HEAD /repos/AprilNEA/OpenLogi/releases/
-    /// latest` request per launch and logs whether a newer version is
-    /// available — no automatic download.
-    #[serde(default)]
-    pub check_for_updates: bool,
-    /// Opt-in automatic install. When true *and* [`Self::check_for_updates`]
-    /// surfaces a newer version, the GUI downloads and stages it in the
-    /// background; the update is applied on the next restart (never mid-session,
-    /// and never auto-relaunched). **Off by default** — it only acts after a
-    /// check the user already opted into, and stays inert in unsigned dev builds
-    /// where verification fails closed.
-    #[serde(default)]
-    pub auto_install_updates: bool,
-    /// True once the first-run "check for updates?" prompt has been answered
-    /// (either way), so it is never shown again. The prompt is how a
-    /// privacy-conscious default of `check_for_updates = false` still lets a
-    /// user opt in on first launch.
-    #[serde(default)]
-    pub update_prompt_seen: bool,
-    /// Whether OpenLogi shows a macOS menu-bar (status item) icon — and, on
-    /// Windows, the notification-area (tray) icon. `true` (default) → the
-    /// agent is visible in the menu bar / tray; `false` → it runs with no
-    /// visible presence (macOS additionally keeps the ordinary Dock icon
-    /// while a window is open). Ignored on Linux.
-    #[serde(default = "default_true")]
-    pub show_in_menu_bar: bool,
-    /// Whether the GUI automatically downloads device images from
-    /// `assets.openlogi.org` when a device appears. `true` (default) keeps
-    /// the current behavior; `false` makes no asset network requests at all
-    /// (the app falls back to bundled art and the synthetic silhouette). A
-    /// manual "Refresh assets" in Settings still fetches on demand regardless.
-    #[serde(default = "default_true")]
-    pub auto_download_assets: bool,
-    /// UI language as a BCP-47-ish locale code matching the GUI's bundled
-    /// locales (e.g. `"en"`, `"de"`, `"pt-BR"`, `"zh-CN"`, `"zh-TW"`; see the
-    /// GUI's `i18n::SUPPORTED`). `None` means "follow the system locale", which
-    /// the GUI resolves at startup. Stored here so a user's explicit choice
-    /// survives restarts regardless of the OS setting.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub language: Option<String>,
-    /// Thumb-wheel responsiveness, on a [`MIN_THUMBWHEEL_SENSITIVITY`]–
-    /// [`MAX_THUMBWHEEL_SENSITIVITY`] scale. It scales both the speed of the
-    /// wheel's continuous horizontal scroll and how few rotation increments a
-    /// custom wheel action needs to fire. [`DEFAULT_THUMBWHEEL_SENSITIVITY`]
-    /// (the out-of-the-box value) means 1× scroll speed; the wheel is only
-    /// diverted from native scrolling once this leaves the default.
-    #[serde(default = "default_thumbwheel_sensitivity")]
-    pub thumbwheel_sensitivity: i32,
-    /// Light/dark appearance preference. Defaults to following the OS.
-    #[serde(default)]
-    pub appearance: Appearance,
-    /// Name of the theme used in light mode (a [`crate`]-agnostic string
-    /// matching a gpui-component theme, e.g. `"OpenLogi Light"`). `None` uses
-    /// the OpenLogi brand light theme.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub theme_light: Option<String>,
-    /// Name of the theme used in dark mode. `None` uses the OpenLogi brand dark
-    /// theme.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub theme_dark: Option<String>,
-    /// Corner-radius override for the UI, in pixels (the Appearance page offers
-    /// `0` / `6` / `12`). `None` keeps each theme's own radius.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub ui_radius: Option<u8>,
-}
-
-/// Out-of-the-box [`AppSettings::thumbwheel_sensitivity`]. At this value the
-/// wheel's horizontal scroll runs at 1× and the wheel is left to scroll
-/// natively (no HID++ diversion) unless a binding diverges from its default.
-pub const DEFAULT_THUMBWHEEL_SENSITIVITY: i32 = 14;
-/// Lowest selectable [`AppSettings::thumbwheel_sensitivity`].
-pub const MIN_THUMBWHEEL_SENSITIVITY: i32 = 1;
-/// Highest selectable [`AppSettings::thumbwheel_sensitivity`].
-pub const MAX_THUMBWHEEL_SENSITIVITY: i32 = 100;
-
-impl AppSettings {
-    /// `skip_serializing_if` helper: true when nothing diverges from the
-    /// default, so empty settings don't clutter `config.toml`.
-    #[must_use]
-    pub fn is_default(&self) -> bool {
-        self == &Self::default()
-    }
-}
-
-impl Default for AppSettings {
-    fn default() -> Self {
-        Self {
-            launch_at_login: false,
-            check_for_updates: false,
-            auto_install_updates: false,
-            update_prompt_seen: false,
-            show_in_menu_bar: true,
-            auto_download_assets: true,
-            language: None,
-            thumbwheel_sensitivity: DEFAULT_THUMBWHEEL_SENSITIVITY,
-            appearance: Appearance::System,
-            theme_light: None,
-            theme_dark: None,
-            ui_radius: None,
-        }
-    }
-}
-
-/// serde default for [`AppSettings::show_in_menu_bar`]: `true`, so the menu-bar
-/// icon is on out of the box and configs predating the field keep that behavior.
-fn default_true() -> bool {
-    true
-}
-
-/// serde default for [`AppSettings::thumbwheel_sensitivity`]: keeps configs
-/// predating the field at the 1× default.
-const fn default_thumbwheel_sensitivity() -> i32 {
-    DEFAULT_THUMBWHEEL_SENSITIVITY
-}
-
-/// Per-device RGB lighting: a single static color, brightness, and on/off.
-/// Deliberately basic — per-key effects are a later addition.
-///
-/// Crosses the agent↔GUI IPC (`set_lighting`), so field order is wire format —
-/// changes require a `PROTOCOL_VERSION` bump (guarded by
-/// `openlogi-agent-core/tests/wire_format.rs`).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Lighting {
-    #[serde(default = "default_lighting_enabled")]
-    pub enabled: bool,
-    /// Static color as 6 hex digits `"RRGGBB"` (no leading `#`).
-    #[serde(default = "default_lighting_color")]
-    pub color: String,
-    /// Brightness percent, clamped to 0–100 on load.
-    #[serde(
-        default = "default_lighting_brightness",
-        deserialize_with = "deserialize_brightness"
-    )]
-    pub brightness: u8,
-}
-
-impl Default for Lighting {
-    fn default() -> Self {
-        Self {
-            enabled: default_lighting_enabled(),
-            color: default_lighting_color(),
-            brightness: default_lighting_brightness(),
-        }
-    }
-}
-
-fn default_lighting_enabled() -> bool {
-    true
-}
-
-fn default_lighting_color() -> String {
-    "ffffff".to_string()
-}
-
-fn default_lighting_brightness() -> u8 {
-    100
-}
-
-/// Clamp a deserialized brightness into the UI's `0..=100` range, so a
-/// hand-edited `config.toml` can't feed out-of-range values into the scaling
-/// math (which assumes `brightness <= 100`).
-fn deserialize_brightness<'de, D>(deserializer: D) -> Result<u8, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    Ok(u8::deserialize(deserializer)?.min(100))
-}
-
-/// Scroll-wheel mode for [`SmartShift`]: free-spin or ratchet (clicky).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum WheelMode {
-    Free,
-    Ratchet,
-}
-
-/// SmartShift auto-disengage out-of-box default (`16` ≈ 4 turn/s, per the
-/// x2110 / x2111 spec). The sensitivity slider's default and the heal target
-/// for a corrupt persisted threshold.
-pub const SMARTSHIFT_AUTO_DISENGAGE_DEFAULT: u8 = 16;
-
-/// Smallest auto-disengage threshold OpenLogi will store or apply (`8` ≈
-/// 2 turn/s). Below this the ratchet releases into free-spin at everyday scroll
-/// speeds, leaving the wheel "stuck" spinning (#317); `0` is also the firmware
-/// "do not change" sentinel that must never be stored as a real value. A
-/// persisted threshold below this floor is a corrupt artifact and is healed to
-/// [`SMARTSHIFT_AUTO_DISENGAGE_DEFAULT`] on load.
-pub const SMARTSHIFT_MIN_AUTO_DISENGAGE: u8 = 8;
-
-/// Heal a persisted auto-disengage threshold on load: anything below
-/// [`SMARTSHIFT_MIN_AUTO_DISENGAGE`] (including the `0` sentinel) becomes the
-/// default. `0xFF` (permanent ratchet) and every real threshold at or above the
-/// floor pass through unchanged.
-fn deserialize_auto_disengage<'de, D>(deserializer: D) -> Result<u8, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let value = u8::deserialize(deserializer)?;
-    Ok(if value < SMARTSHIFT_MIN_AUTO_DISENGAGE {
-        tracing::warn!(
-            value,
-            min = SMARTSHIFT_MIN_AUTO_DISENGAGE,
-            default = SMARTSHIFT_AUTO_DISENGAGE_DEFAULT,
-            "healed persisted SmartShift auto-disengage threshold below supported floor"
-        );
-        SMARTSHIFT_AUTO_DISENGAGE_DEFAULT
-    } else {
-        value
-    })
-}
-
-/// Per-device SmartShift wheel configuration, persisted so the agent can
-/// re-apply it when the device reconnects: the values are written to device
-/// RAM and do not survive a power cycle (#189), despite earlier assumptions
-/// that the device kept them in NVM.
-///
-/// Config-file only — never crosses the IPC (the agent reads it from
-/// `config.toml` on reload), so it is free to evolve without a
-/// `PROTOCOL_VERSION` bump.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SmartShift {
-    pub mode: WheelMode,
-    /// SmartShift auto-disengage threshold (`0x08`–`0xFE`, in 0.25 turn/s
-    /// steps), or `0xFF` for a permanently engaged ratchet. A persisted value
-    /// below [`SMARTSHIFT_MIN_AUTO_DISENGAGE`] is healed to the default on load.
-    #[serde(deserialize_with = "deserialize_auto_disengage")]
-    pub auto_disengage: u8,
-    /// Tunable-torque force percentage (`1`–`100`), `0` when the device
-    /// doesn't support tunable torque.
-    pub tunable_torque: u8,
-}
-
-/// Which control owns a device's single gesture role.
-///
-/// Stored explicitly — rather than inferred from which button happens to carry a
-/// [`Binding::Gesture`] — so switching the gesture button never has to collapse
-/// a button's gesture map to encode the choice: every gesture-capable button
-/// keeps its full direction map, and only the owner is dispatched. Serialized as
-/// a bare string (`"Off"` or a [`ButtonId`] name) so it stays a TOML scalar.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum GestureOwner {
-    /// Gestures are explicitly turned off for this device.
-    Off,
-    /// The named button owns the gesture role.
-    Button(ButtonId),
-}
-
-impl Serialize for GestureOwner {
-    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        match self {
-            // "Off" can't collide with a ButtonId variant name (all CamelCase
-            // control names), so the string space is unambiguous.
-            GestureOwner::Off => serializer.serialize_str("Off"),
-            GestureOwner::Button(id) => id.serialize(serializer),
-        }
-    }
-}
-
-/// Lenient field deserializer for [`RawDeviceConfig::gesture_owner`]. An
-/// unrecognized or miscased value (`"back"`, a typo, a future-version button
-/// name) is treated as absent — i.e. "infer the owner" — rather than failing the
-/// whole-document parse and reverting *every* device's settings to defaults.
-/// Mirrors [`deserialize_brightness`], which clamps a bad value instead of
-/// erroring; a hand-editable config should degrade one field, not the document.
-fn deserialize_gesture_owner<'de, D>(deserializer: D) -> Result<Option<GestureOwner>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let s = String::deserialize(deserializer)?;
-    if s == "Off" {
-        return Ok(Some(GestureOwner::Off));
-    }
-    // Parse the button name with a throwaway error type so an unknown token maps
-    // to `None` (infer) rather than propagating an error.
-    let button = ButtonId::deserialize(
-        serde::de::value::StrDeserializer::<serde::de::value::Error>::new(&s),
-    )
-    .ok();
-    Ok(button.map(GestureOwner::Button))
-}
-
-/// Last-known identity of a device, captured while it was online so the UI can
-/// render its card and the *correct* config panels before any live HID++ probe
-/// completes — or while the device is asleep and can't be probed at all.
-///
-/// Every field is a **static property of the model**, not of the current
-/// connection: an MX Master 3S has adjustable DPI whether or not it is awake.
-/// That is what makes this safe to persist — it never goes stale. It is also
-/// free of any per-unit identifier (no serial number, no unit id), so caching
-/// it adds no privacy surface beyond the `config_key` already used as the map
-/// key. Persisting identity is what stops a sleeping/just-booted mouse from
-/// vanishing from the device list (and losing its Pointer/Buttons panels)
-/// until a cold probe happens to win its race — see issue #159.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct DeviceIdentity {
-    /// The name shown in the carousel, as resolved from the asset registry the
-    /// last time the device was online.
-    pub display_name: String,
-    /// HID++ model identity from feature 0x0003, when available. Persisted so
-    /// the GUI can resolve the same curated asset while the device is asleep.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub model_info: Option<DeviceModelInfo>,
-    /// Firmware codename, when available. Used as an asset-resolution hint and
-    /// as a readable fallback for devices without curated model metadata.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub codename: Option<String>,
-    /// The device's resolved [`DeviceKind`] (asset registry preferred, HID++
-    /// classification as fallback).
-    pub kind: DeviceKind,
-    /// Configuration capabilities measured from the device's HID++ feature
-    /// table. This is the field that keeps a sleeping mouse's panels visible.
-    pub capabilities: Capabilities,
-}
-
-/// Settings scoped to a single physical device.
-///
-/// Deserialization goes through `RawDeviceConfig` (`#[serde(from)]`) so
-/// pre-v2 files — which split bindings across `button_bindings` +
-/// `gesture_bindings` — fold into the unified [`Self::bindings`] map. Only
-/// `bindings` is ever serialized, so a migrated file self-heals to the v2 shape
-/// on its next save.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(from = "RawDeviceConfig")]
-pub struct DeviceConfig {
-    /// Which button owns the device's single gesture role, once the user has
-    /// chosen explicitly. Absent means "infer" (the dedicated HID++ gesture
-    /// button owns gestures if present) — see [`Config::gesture_owner`]. Listed
-    /// first so it serializes as a scalar ahead of the `bindings` sub-table.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub gesture_owner: Option<GestureOwner>,
-    /// Last-known identity (name / kind / capabilities), captured while the
-    /// device was online. Lets the UI render this device — with the right
-    /// config panels — on a cold start before any probe, or while it sleeps.
-    /// `None` for configs written before this field existed or by hand.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub identity: Option<DeviceIdentity>,
-    /// Every rebindable button's binding: a single [`Action`], or — for the
-    /// gesture button (and, later, any raw-XY-capable button) — a
-    /// [`Binding::Gesture`] per-direction map.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub bindings: BTreeMap<ButtonId, Binding>,
-    /// Per-application binding overlays (P1.4). Keyed by bundle identifier
-    /// (e.g. `"com.microsoft.VSCode"` on macOS). When the foreground app's
-    /// id matches a key here, those bindings take precedence; anything not
-    /// listed falls through to `bindings`. Deliberately `Action`-valued (not
-    /// `Binding`): a per-app override replaces the whole button with one
-    /// action, never a per-direction gesture overlay.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub per_app_bindings: BTreeMap<String, BTreeMap<ButtonId, Action>>,
-    /// Ordered list of DPI presets cycled through by
-    /// [`Action::CycleDpiPresets`] and indexed by
-    /// [`Action::SetDpiPreset`]. Empty means "no presets configured" —
-    /// the cycle action becomes a no-op until the user adds at least one.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub dpi_presets: Vec<u32>,
-    /// The sensor DPI the user committed for this device. Persisted because
-    /// the value lives in device RAM and resets on a power cycle (#189); the
-    /// agent re-applies it when the device reconnects. `None` until the user
-    /// first changes DPI.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub dpi: Option<u32>,
-    /// Per-device RGB lighting (static color + brightness + on/off). `None`
-    /// until the user changes it, so it stays out of `config.toml` otherwise.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub lighting: Option<Lighting>,
-    /// Per-device SmartShift wheel configuration, re-applied on reconnect for
-    /// the same reason as [`Self::dpi`]. `None` until the user changes it.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub smartshift: Option<SmartShift>,
-    /// Invert this device's scroll-wheel direction relative to the OS setting
-    /// (issue #126): on, a wheel tick scrolls the opposite way, so a user who
-    /// keeps macOS "natural scrolling" for the trackpad can have a traditional
-    /// "reverse" wheel on the mouse. Vertical only; the agent applies it through
-    /// the device's HID++ native wheel-inversion mode when supported. `false`
-    /// (default) is the native direction, and is omitted from `config.toml`.
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub invert_scroll: bool,
-}
-
-/// `skip_serializing_if` helper for plain `bool` fields whose default is
-/// `false`: keeps an unset toggle out of `config.toml` entirely.
-#[allow(
-    clippy::trivially_copy_pass_by_ref,
-    reason = "serde's skip_serializing_if requires a fn(&T) -> bool signature"
-)]
-fn is_false(b: &bool) -> bool {
-    !*b
-}
-
-/// Deserialize-only shim that folds the pre-v2 `button_bindings` +
-/// `gesture_bindings` fields into [`DeviceConfig::bindings`]. Never serialized
-/// (only [`DeviceConfig`] is), so reading a legacy file and saving rewrites it
-/// in the v2 shape.
-#[derive(Deserialize)]
-struct RawDeviceConfig {
-    /// Explicit gesture owner (v2.1+). Absent on older configs → `None` → the
-    /// owner is inferred in [`Config::gesture_owner`]. A present-but-invalid
-    /// value is tolerated as `None` (infer), not a parse error — see
-    /// [`deserialize_gesture_owner`].
-    #[serde(default, deserialize_with = "deserialize_gesture_owner")]
-    gesture_owner: Option<GestureOwner>,
-    #[serde(default)]
-    identity: Option<DeviceIdentity>,
-    /// v2 shape — present on already-migrated files; wins on any key collision.
-    #[serde(default)]
-    bindings: BTreeMap<ButtonId, Binding>,
-    /// Legacy v1 per-button single bindings.
-    #[serde(default)]
-    button_bindings: BTreeMap<ButtonId, Action>,
-    /// Legacy v1 flat gesture map (implicitly the gesture button's directions).
-    #[serde(default)]
-    gesture_bindings: BTreeMap<GestureDirection, Action>,
-    #[serde(default)]
-    per_app_bindings: BTreeMap<String, BTreeMap<ButtonId, Action>>,
-    #[serde(default)]
-    dpi_presets: Vec<u32>,
-    #[serde(default)]
-    dpi: Option<u32>,
-    #[serde(default)]
-    lighting: Option<Lighting>,
-    #[serde(default)]
-    smartshift: Option<SmartShift>,
-    #[serde(default)]
-    invert_scroll: bool,
-}
-
-impl From<RawDeviceConfig> for DeviceConfig {
-    fn from(raw: RawDeviceConfig) -> Self {
-        let mut bindings = raw.bindings; // the v2 map wins on every key.
-
-        // Re-home the legacy flat gesture map under `GestureButton`. This MUST
-        // happen before folding `button_bindings`, so a legacy single
-        // `button_bindings[GestureButton]` entry coexisting with a
-        // `gesture_bindings` map cannot claim the slot first and silently drop
-        // the whole direction map (the pre-v2 rule was "gesture entries win").
-        if !raw.gesture_bindings.is_empty() {
-            bindings
-                .entry(ButtonId::GestureButton)
-                .or_insert_with(|| Binding::Gesture(raw.gesture_bindings));
-        }
-        for (button, action) in raw.button_bindings {
-            // A legacy `button_bindings[GestureButton]` is vestigial and must not
-            // become a `Binding::Single`: the gesture button never dispatched
-            // through the per-button map (it is not an OS-hook button, and its
-            // plain press routes through the gesture `Click` slot — see
-            // agent-core `bindings_for`). A `Single` here would be unreachable —
-            // the GUI hides it and the runtime ignores it — while folding it into
-            // `Click` would resurrect a dead binding as a behavior change. Drop
-            // it: the gesture map (re-homed above) already owns this button, and
-            // an absent entry falls back to the canonical default, exactly as
-            // pre-v2.
-            if button == ButtonId::GestureButton {
-                continue;
-            }
-            bindings.entry(button).or_insert(Binding::Single(action));
-        }
-
-        DeviceConfig {
-            gesture_owner: raw.gesture_owner,
-            identity: raw.identity,
-            bindings,
-            per_app_bindings: raw.per_app_bindings,
-            dpi_presets: raw.dpi_presets,
-            dpi: raw.dpi,
-            lighting: raw.lighting,
-            smartshift: raw.smartshift,
-            invert_scroll: raw.invert_scroll,
         }
     }
 }
@@ -1137,32 +640,6 @@ mod tests {
             !body.contains("invert_scroll"),
             "default invert_scroll should be omitted: {body}"
         );
-    }
-
-    #[test]
-    fn low_auto_disengage_heals_to_default_on_load() {
-        // A pre-#317 config could persist a runaway-low threshold (or the `0`
-        // sentinel); loading it must heal to the default so reapply doesn't
-        // re-program free-spin-on-any-scroll into the device — while a real
-        // threshold and the `0xFF` permanent-ratchet value pass through.
-        let heal = |v: u8| {
-            let body = format!("mode = \"ratchet\"\nauto_disengage = {v}\ntunable_torque = 50\n");
-            toml::from_str::<SmartShift>(&body)
-                .expect("parse")
-                .auto_disengage
-        };
-        assert_eq!(heal(0), SMARTSHIFT_AUTO_DISENGAGE_DEFAULT);
-        assert_eq!(heal(1), SMARTSHIFT_AUTO_DISENGAGE_DEFAULT);
-        assert_eq!(
-            heal(SMARTSHIFT_MIN_AUTO_DISENGAGE - 1),
-            SMARTSHIFT_AUTO_DISENGAGE_DEFAULT
-        );
-        assert_eq!(
-            heal(SMARTSHIFT_MIN_AUTO_DISENGAGE),
-            SMARTSHIFT_MIN_AUTO_DISENGAGE
-        );
-        assert_eq!(heal(16), 16);
-        assert_eq!(heal(0xff), 0xff);
     }
 
     #[test]
